Coneccion Del Front con el Backend de las propiedades

+++ visita-facil-api/src/main/java/com/construsoft/visita_facil_api/service/PropiedadService.java
@@ -22,14 +22,15 @@ public class PropiedadService {
     private PropiedadRepository propiedadRepository;
 
     public Optional<Propiedad> getById(int id) {
-        return propiedadRepository.findById((long) id);
+        return propiedadRepository.findById(id);
     }
 
     public List<Propiedad> getPropiedadesDisponibles() {
         return propiedadRepository.findByDisponibleTrue();
     }
 
-    public Propiedad addPropiedad(String titulo, String descripcion, int precio, String tipo, String ubicacion, boolean disponible) {
+    public Propiedad addPropiedad(String titulo, String descripcion, int precio, String tipo, String ubicacion,
+            boolean disponible) {
         Propiedad propiedad = new Propiedad();
         propiedad.setTitulo(titulo);
         propiedad.setDescripcion(descripcion);
@@ -41,7 +42,7 @@ public class PropiedadService {
     }
 
     public List<String> subirImagenes(int id, List<MultipartFile> imagenes) throws IOException {
-        Propiedad propiedad = propiedadRepository.findById((long) id)
+        Propiedad propiedad = propiedadRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("Propiedad no encontrada"));
 
         List<String> urls = new ArrayList<>();
